Create and try using CachedHashableStateFactory for RMAXQ
