+++ config/cleanup/rmaxqTest.yaml
@@ -15,12 +15,12 @@ rewardNoop: -1.0
 rewardPull: 0.0
 
 agents:
-  - ramdp
   - rmaxq
+  - ramdp
 
-episodes: 60
-max_steps: 1000
-trials: 20
+episodes: 50
+max_steps: 500
+trials: 4
 gamma: 0.95
 
 planning:
@@ -28,7 +28,7 @@ planning:
 
 rmax:
   vmax: 999
-  threshold: 1
+  threshold: 3
   max_delta: 0.0001
 
 output:
